Split tutorial patch classes
